requirements.txt
aiohttp==3.7.4.post0
arrow==0.14.7
async-timeout==3.0.1
attrs==20.3.0
certifi==2020.12.5
chardet==4.0.0
discord.py==1.7.2
ics==0.7
idna==2.10
multidict==5.1.0
python-dateutil==2.8.1
python-dotenv==0.17.1
requests==2.25.1
six==1.15.0
TatSu==5.6.1
typing-extensions==3.10.0.0
urllib3==1.26.4
yarl==1.6.3
